lista 2 exercicios

+++ capitulo 5/lista de exercicios/exercicio5.py
@@ -2,6 +2,6 @@ def calcula_peso (altura):
     pesoIdeal = (62.1 * altura) - 44.7
     print(pesoIdeal)
 
-altura = int(input("Escreva sua altura: "))
+altura = float(input("Escreva sua altura: "))
 
 calcula_peso(altura)
